requirements.txt
aiocontextvars==0.2.2
apex==0.1
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.3
contextvars==2.4
dataclasses==0.8
Flask==1.1.2
glob2==0.7
idna==2.10
immutables==0.15
itsdangerous==1.1.0
Jinja2==2.11.3
loguru==0.5.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
opencv-python==4.5.1.48
pathspec==0.8.1
Pillow==8.1.2
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.2
torch==1.8.0+cu101
torchtext==0.4.0
torchvision==0.9.0+cu101
tqdm==4.59.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
